Project Types (OpenBots, Python, TagUI, CSScript)

+++ OpenBots.Core/Project/Project.cs
@@ -2,6 +2,7 @@
 using ICSharpCode.SharpZipLib.Zip;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using OpenBots.Core.Enums;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -16,6 +17,7 @@ namespace OpenBots.Core.Project
     {
         public Guid ProjectID { get; set; }
         public string ProjectName { get; set; }
+        public ProjectType ProjectType { get; set; }
         public string Main { get; set; }
         public string Version { get; set; }
         public Dictionary<string, string> Dependencies { get; set; }
@@ -30,15 +32,34 @@ namespace OpenBots.Core.Project
             "UIAutomation"
         };
 
-        public Project(string projectName)
+        public Project(string projectName, ProjectType projectType)
         {
             ProjectID = Guid.NewGuid();
             ProjectName = projectName;
-            Main = "Main.json";
             Version = Application.ProductVersion;
+            ProjectType = projectType;
 
             var commandVersion = Regex.Matches(Application.ProductVersion, @"\d+\.\d+\.\d+")[0].ToString();
-            Dependencies = DefaultCommandGroups.ToDictionary(x => $"OpenBots.Commands.{x}", x => commandVersion);
+            
+            switch (ProjectType)
+            {
+                case ProjectType.OpenBots:
+                    Main = "Main.obscript";
+                    Dependencies = DefaultCommandGroups.ToDictionary(x => $"OpenBots.Commands.{x}", x => commandVersion);
+                    break;
+                case ProjectType.Python:
+                    Main = "Main.py";
+                    Dependencies = new Dictionary<string, string>();
+                    break;
+                case ProjectType.TagUI:
+                    Main = "Main.tag";
+                    Dependencies = new Dictionary<string, string>();
+                    break;
+                case ProjectType.CSScript:
+                    Main = "Main.cs";
+                    Dependencies = new Dictionary<string, string>();
+                    break;
+            }
         }
 
         public void SaveProject(string scriptPath)
@@ -55,11 +76,11 @@ namespace OpenBots.Core.Project
                     projectPath = Path.GetDirectoryName(scriptPath);
                     DirectoryInfo dirInfo = new DirectoryInfo(projectPath);
                     dirName = dirInfo.Name;
-                    configPath = Path.Combine(projectPath, "project.config");
+                    configPath = Path.Combine(projectPath, "project.obconfig");
                     scriptPath = projectPath;
                 } while (dirName != ProjectName || !File.Exists(configPath));
 
-                //If requirements are met, a project.config is created/updated
+                //If requirements are met, a project.obconfig is created/updated
                 if (dirName == ProjectName && File.Exists(configPath))
                 {
                     Version = Application.ProductVersion;
@@ -74,51 +95,55 @@ namespace OpenBots.Core.Project
 
         public static void RenameProject(Project newProject, string newProjectPath)
         {
-            string configPath = Path.Combine(newProjectPath, "project.config");
+            string configPath = Path.Combine(newProjectPath, "project.obconfig");
 
             if (File.Exists(configPath))
                 File.WriteAllText(configPath, JsonConvert.SerializeObject(newProject));
             else
-                throw new FileNotFoundException("project.config not found. Unable to save project.");
+                throw new FileNotFoundException("project.obconfig not found. Unable to save project.");
         }
 
         public static Project OpenProject(string configFilePath)
         {
-            //Loads project from project.config
+            //Loads project from project.obconfig
             if (File.Exists(configFilePath))
             {
                 string projectJSONString = File.ReadAllText(configFilePath);
                 var project = JsonConvert.DeserializeObject<Project>(projectJSONString);
-                string dialogMessageFirstLine = "";
 
-                if (!projectJSONString.Contains("Version"))
+                if (project.ProjectType == ProjectType.OpenBots)
                 {
-                    dialogMessageFirstLine = $"Attempting to open a 'project.config' from a version of OpenBots Studio older than 1.2.0.0.";
-                }
-                //if project version is lower than than 1.3.0.0
-                else if (new Version(project.Version).CompareTo(new Version("1.4.0.0")) < 0)
-                {
-                    dialogMessageFirstLine = $"Attempting to open a 'project.config' from OpenBots Studio version {project.Version}.";
+                    string dialogMessageFirstLine = "";
 
-                    var dialogResult = MessageBox.Show($"{dialogMessageFirstLine} " +
-                                                   $"Would you like to attempt to convert this config file to {Application.ProductVersion}? " +
-                                                   "\n\nWarning: Once a 'project.config' has been converted, it cannot be undone.",
-                                                   "Convert 'project.config'", MessageBoxButtons.YesNo);
-
-                    if (dialogResult == DialogResult.Yes)
+                    if (!projectJSONString.Contains("Version"))
                     {
-                        project.Version = Application.ProductVersion;
-                        var commandVersion = Regex.Matches(Application.ProductVersion, @"\d+\.\d+\.\d+")[0].ToString();
-                        project.Dependencies = DefaultCommandGroups.ToDictionary(x => $"OpenBots.Commands.{x}", x => commandVersion);
-                        File.WriteAllText(configFilePath, JsonConvert.SerializeObject(project));
-                    }                        
+                        dialogMessageFirstLine = $"Attempting to open a 'project.obconfig' from a version of OpenBots Studio older than 1.2.0.0.";
+                    }
+                    //if project version is lower than than 1.3.0.0
+                    else if (new Version(project.Version).CompareTo(new Version("1.4.0.0")) < 0)
+                    {
+                        dialogMessageFirstLine = $"Attempting to open a 'project.obconfig' from OpenBots Studio version {project.Version}.";
+
+                        var dialogResult = MessageBox.Show($"{dialogMessageFirstLine} " +
+                                                       $"Would you like to attempt to convert this config file to {Application.ProductVersion}? " +
+                                                       "\n\nWarning: Once a 'project.obconfig' has been converted, it cannot be undone.",
+                                                       "Convert 'project.obconfig'", MessageBoxButtons.YesNo);
+
+                        if (dialogResult == DialogResult.Yes)
+                        {
+                            project.Version = Application.ProductVersion;
+                            var commandVersion = Regex.Matches(Application.ProductVersion, @"\d+\.\d+\.\d+")[0].ToString();
+                            project.Dependencies = DefaultCommandGroups.ToDictionary(x => $"OpenBots.Commands.{x}", x => commandVersion);
+                            File.WriteAllText(configFilePath, JsonConvert.SerializeObject(project));
+                        }
+                    }
                 }
-                    
+                                  
                 return project;
             }
             else
             {
-                throw new Exception("project.config Not Found");
+                throw new Exception("project.obconfig Not Found");
             }
         }
 
@@ -141,7 +166,7 @@ namespace OpenBots.Core.Project
             File.Delete(processNugetFilePath);
 
             //get config file and rename project
-            string configFilePath = Directory.GetFiles(projectDirectory, "project.config", SearchOption.AllDirectories).First();
+            string configFilePath = Directory.GetFiles(projectDirectory, "project.obconfig", SearchOption.AllDirectories).First();
             var config = JObject.Parse(File.ReadAllText(configFilePath));
             config["ProjectName"] = new DirectoryInfo(projectDirectory).Name;
             File.WriteAllText(configFilePath, JsonConvert.SerializeObject(config));
